chore(functions): Refactor functions error logic
