+++ src/functions/index.ts
@@ -72,4 +72,4 @@ export function getFunctions(app?: App): Functions {
   return firebaseApp.getOrInitService('functions', (app) => new Functions(app));
 }
 
-export { FirebaseFunctionsError, FunctionsErrorCode } from './functions-api-client-internal';
+export { FirebaseFunctionsError, FunctionsErrorCode } from './error';
